requirements.txt
# Core dependencies
asyncio-mqtt==0.16.1
aiohttp==3.11.11
aiofiles==24.1.0
aioredis==2.0.1
asyncpg==0.29.0
sqlalchemy==1.4.24
alembic==1.13.1

# Data processing
pandas==2.1.4
numpy==1.26.4
pyarrow==14.0.1
fastparquet==2023.10.1

# Web scraping
lxml==5.3.0
requests==2.31.0
urllib3==2.1.0
crawl4ai==0.7.4
cryptography==42.0.8
litellm==1.75.5.post1
playwright==1.49.0

# API and web framework
fastapi==0.115.6
uvicorn==0.24.0
pydantic==2.11.7
pydantic-settings==2.10.1

# Task scheduling
celery==5.3.4
redis==5.0.1
croniter==2.0.1

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0
jsonschema==4.22.0

# Logging and monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0

# Security
cryptography>=42.0.0,<47
passlib==1.7.4
python-jose==3.3.0

# Database drivers
psycopg2-binary==2.9.9
pymongo==4.6.0
motor==3.3.2
clickhouse-driver==0.2.6
clickhouse-sqlalchemy==0.2.4

# Message queues
pika==1.3.2
kombu==5.3.4

# Time and date
python-dateutil==2.8.2
pytz==2023.3

# HTTP clients
httpx==0.27.2
aiohttp-client-cache==0.8.0

# Data validation
marshmallow==3.20.1
cerberus==1.3.5

# Caching
diskcache==5.6.3
aiocache==0.12.2

# File handling
openpyxl==3.1.2
xlsxwriter==3.1.9

# Image processing (if needed)
Pillow==10.4.0

# Machine learning (optional)

# Natural language processing
nltk==3.9.1

# Financial data
yfinance==0.2.28
alpha-vantage==2.3.1
quandl==3.7.0

# WebSocket support
websockets==12.0

# Background tasks
apscheduler==3.10.4

# Metrics and monitoring
psutil==6.1.1
memory-profiler==0.61.0

# Serialization
msgpack==1.0.7
orjson==3.9.10

# Environment management
python-decouple==3.8

# Type checking
types-requests==2.31.0.10
types-PyYAML==6.0.12.12
